requirements.txt
dash==3.0.4
dash-leaflet==1.0.0
pandas==2.2.3
plotly==6.0.1
dash-bootstrap-components==2.0.2
google-cloud-bigquery==3.32.0
google-cloud-storage==3.1.0
gunicorn==20.1.0
db-dtypes==1.4.3